rename ProgramInfo to CompilationUnit
CompilationUnit is now honestly per-compilation: rich AST nodes for the
entry file plus the scoped type-alias map. Cross-file work (imported
type definitions, imported safe markings, imported function signatures)
flows through a single SymbolTable.resolveImport pass instead of three
separate hand-rolled loops.

The importedFunctions field is kept (typescriptBuilder and lsp/completion
still consume it) but its population is now a clean derivation: one loop
over import statements, calling resolveImport, populating both
importedFunctions and safeFunctions and merged type aliases together.

CLAUDE.md updated to reflect the new pipeline names and the SymbolTable
class API.
