+++ packages/agency-lang/lib/backends/typescriptGenerator.ts
@@ -1,6 +1,6 @@
 import { AgencyProgram } from "../types.js";
 import { TypescriptPreprocessor } from "@/preprocessors/typescriptPreprocessor.js";
-import { collectProgramInfo, type ProgramInfo } from "@/programInfo.js";
+import { buildCompilationUnit, type CompilationUnit } from "@/compilationUnit.js";
 import { AgencyConfig } from "@/config.js";
 import { TypeScriptBuilder } from "./typescriptBuilder.js";
 import { printTs } from "../ir/prettyPrint.js";
@@ -8,14 +8,14 @@ import { printTs } from "../ir/prettyPrint.js";
 export function generateTypeScript(
   program: AgencyProgram,
   config?: AgencyConfig,
-  info?: ProgramInfo,
+  info?: CompilationUnit,
   moduleId?: string,
   outputFile?: string,
 ): string {
   if (!moduleId) {
     throw new Error("moduleId is required for generateTypeScript");
   }
-  const programInfo = info ?? collectProgramInfo(program);
+  const programInfo = info ?? buildCompilationUnit(program);
   const preprocessor = new TypescriptPreprocessor(program, config, programInfo);
   const preprocessedProgram = preprocessor.preprocess();
   const builder = new TypeScriptBuilder(config, programInfo, moduleId, outputFile);
