+++ packages/agency-lang/lib/lsp/completion.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect } from "vitest";
 import { CompletionItemKind } from "vscode-languageserver-protocol";
 import { getCompletions } from "./completion.js";
 import { parseAgency } from "../parser.js";
-import { collectProgramInfo } from "../programInfo.js";
+import { buildCompilationUnit } from "../compilationUnit.js";
 import { SymbolTable } from "@/symbolTable.js";
 
 function parse(source: string) {
@@ -14,7 +14,7 @@ function parse(source: string) {
 describe("getCompletions", () => {
   it("includes function definitions as Function items", () => {
     const program = parse("def greet() { let x: string = `hi` }");
-    const info = collectProgramInfo(program, new SymbolTable());
+    const info = buildCompilationUnit(program, new SymbolTable());
     const items = getCompletions(info);
     const greet = items.find((i) => i.label === "greet");
     expect(greet).toBeDefined();
@@ -23,7 +23,7 @@ describe("getCompletions", () => {
 
   it("includes graph nodes as Module items", () => {
     const program = parse("node main() { let x: number = 1 }");
-    const info = collectProgramInfo(program, new SymbolTable());
+    const info = buildCompilationUnit(program, new SymbolTable());
     const items = getCompletions(info);
     const main = items.find((i) => i.label === "main");
     expect(main).toBeDefined();
@@ -32,7 +32,7 @@ describe("getCompletions", () => {
 
   it("includes type aliases as TypeParameter items", () => {
     const program = parse("type Name = string\ndef use(n: Name) { }");
-    const info = collectProgramInfo(program, new SymbolTable());
+    const info = buildCompilationUnit(program, new SymbolTable());
     const items = getCompletions(info);
     const name = items.find((i) => i.label === "Name");
     expect(name).toBeDefined();
@@ -41,7 +41,7 @@ describe("getCompletions", () => {
 
   it("deduplicates items", () => {
     const program = parse("def foo() { } \ndef foo2() { }");
-    const info = collectProgramInfo(program, new SymbolTable());
+    const info = buildCompilationUnit(program, new SymbolTable());
     const items = getCompletions(info);
     const labels = items.map((i) => i.label);
     const unique = new Set(labels);
@@ -49,7 +49,7 @@ describe("getCompletions", () => {
   });
 
   it("returns empty array for empty program", () => {
-    const info = collectProgramInfo(
+    const info = buildCompilationUnit(
       { type: "agencyProgram", nodes: [] },
       new SymbolTable(),
     );
